Moved filter and sort to helpers folder
